requirements.txt
asyncio-redis==0.16.0
redis==5.2.1
aiobotocore
boto3
PyJWT==2.8.0
jwcrypto==1.5.1
requests==2.32.3
pytz==2024.2
pydantic==2.10.5
fastapi==0.115.6
python-decouple==3.8
fastapi==0.115.6
mongoengine==0.29.1